+++ test/parse.spec.js
@@ -9,15 +9,15 @@
         helper = require('../src/lib/helper');
 
     describe('Parser Suite', function () {
-        var options;
+        var options,
+            JS_SOURCE = '/**\n *  describe\n *  @namespace\n *  @type {Object}\n */\nconst nspace = {};\nclass Test { constructor() {} }';
 
         // beforeAll(function () {});
 
         it('should throw on invalid source file(s)', function (done) {
-            console.log('options', options);
             var throwTest = function () {
                 try {
-                    jsdocx.parse(options).finally(done);
+                    jsdocx.parse(undefined).finally(done);
                 } catch (e) {
                     done();
                 }
@@ -86,7 +86,9 @@
 
         it('should parse multiple files', function (done) {
             options = {
-                files: ['./test/input/code.es6.js', './test/input/test2.es6.js'],
+                files: [
+                    './test/input/**/*.js'
+                ],
                 encoding: 'utf8',
                 recurse: false,
                 pedantic: false,
@@ -96,14 +98,21 @@
                 undocumented: true,
                 undescribed: true,
                 relativePath: null,
-                filter: null
+                filter: null,
+                output: {
+                    path: './test/output/docs-multiple-files.json',
+                    indent: true
+                }
             };
             jsdocx.parse(options)
                 .then(function (docs) {
                     expect(docs).toEqual(jasmine.any(Array));
-                    var result = _.filter(docs, { meta: { filename: 'code.es6.js' } });
+                    var result;
+                    result = _.filter(docs, { meta: { filename: 'code.es6.js' } });
                     expect(result.length).toBeGreaterThan(0);
                     result = _.filter(docs, { meta: { filename: 'test2.es6.js' } });
+                    expect(result.length).toBeGreaterThan(0);
+                    result = _.filter(docs, { meta: { filename: 'test3.es5.js' } });
                     expect(result.length).toBeGreaterThan(0);
                 })
                 .catch(function (err) {
@@ -113,9 +122,159 @@
                 .finally(done);
         });
 
-        it('should parse source code', function (done) {
+        it('should sort symbols (ungrouped)', function (done) {
+            options.sort = true;
+            // options.output = false;
+            jsdocx.parse(options)
+                .then(function (docs) {
+                    expect(docs).toEqual(jasmine.any(Array));
+
+                    var cache = {};
+                    function index(longname) {
+                        cache[longname] = cache[longname]
+                            || _.findIndex(docs, { longname: longname });
+                        return cache[longname];
+                    }
+
+                    expect(index('_opt')).toBeLessThan(index('<anonymous>~namespace'));
+                    expect(index('aOption')).toBeLessThan(index('Code'));
+                    expect(index('Code')).toBeLessThan(index('Code#_'));
+                    expect(index('Code')).toBeLessThan(index('module:test2'));
+                    expect(index('Code')).toBeLessThan(index('namespace'));
+                    expect(index('module:test2')).toBeLessThan(index('namespace'));
+                    expect(index('namespace')).toBeLessThan(index('namespace.location'));
+                    expect(index('aOption')).toBeLessThan(index('zOption'));
+                    expect(index('aOption')).toBeLessThan(index('bOption'));
+                    expect(index('bOption')).toBeLessThan(index('zOption'));
+                    expect(index('Code.aStaticMethod')).toBeLessThan(index('Code#bInstanceMethod'));
+                    expect(index('Code#bInstanceMethod')).toBeLessThan(index('Code.xStaticMethod'));
+                    expect(index('Code.utils')).toBeLessThan(index('Code.xStaticMethod'));
+                    expect(index('Code#instanceMethod')).toBeLessThan(index('Code.utils'));
+
+                    // docs.forEach(function (symbol) {
+                    //     console.log(symbol.longname);
+                    //     if (Array.isArray(symbol.properties)) {
+                    //         symbol.properties.forEach(function (prop) {
+                    //             console.log('prop:', prop.name);
+                    //         });
+                    //     }
+                    // });
+                    // console.log('-------------------');
+                })
+                .catch(function (err) {
+                    expect(Boolean(err)).toEqual(false);
+                    console.log(err.stack || err);
+                })
+                .finally(done);
+        });
+
+        it('should sort symbols (grouped)', function (done) {
+            options.sort = 'grouped';
+            // options.output = false;
+            jsdocx.parse(options)
+                .then(function (docs) {
+                    expect(docs).toEqual(jasmine.any(Array));
+
+                    var cache = {};
+                    function index(longname) {
+                        cache[longname] = cache[longname]
+                            || _.findIndex(docs, { longname: longname });
+                        return cache[longname];
+                    }
+
+                    expect(index('_opt')).toBeLessThan(index('<anonymous>~namespace'));
+                    expect(index('aOption')).toBeLessThan(index('Code'));
+                    expect(index('Code')).toBeLessThan(index('Code#_'));
+                    expect(index('Code')).toBeLessThan(index('module:test2'));
+                    expect(index('Code')).toBeLessThan(index('namespace'));
+                    expect(index('module:test2')).toBeLessThan(index('namespace'));
+                    expect(index('namespace')).toBeLessThan(index('namespace.location'));
+                    expect(index('aOption')).toBeLessThan(index('zOption'));
+                    expect(index('aOption')).toBeLessThan(index('bOption'));
+                    expect(index('bOption')).toBeLessThan(index('zOption'));
+                    expect(index('Code.aStaticMethod')).toBeLessThan(index('Code#bInstanceMethod'));
+                    expect(index('Code#bInstanceMethod')).not.toBeLessThan(index('Code.xStaticMethod'));
+                    expect(index('Code#instanceMethod')).not.toBeLessThan(index('Code.utils'));
+
+                    // docs.forEach(function (item) {
+                    //     console.log(item.longname);
+                    // });
+                })
+                .catch(function (err) {
+                    expect(Boolean(err)).toEqual(false);
+                    console.log(err.stack || err);
+                })
+                .finally(done);
+        });
+
+        it('should build hierarchical symbols (grouped)', function (done) {
+            options.files = [
+                './test/input/code.es6.js',
+                './test/input/test3.es5.js'
+            ];
+            options.sort = 'grouped';
+            options.hierarchy = true;
+            options.module = false;
+            options.undocumented = false;
+            options.undescribed = false;
+            options.output = {
+                path: './test/output/docs-hierarchy.json',
+                indent: true
+            };
+            jsdocx.parse(options)
+                .then(function (docs) {
+                    expect(docs).toEqual(jasmine.any(Array));
+                    expect(docs.length).toEqual(2);
+
+                    function getIndexer(symbol, prop, nameProp) {
+                        nameProp = nameProp || 'longname';
+                        var cache = {};
+                        return function index(val) {
+                            cache[val] = cache[val]
+                                || _.findIndex(symbol[prop], function (s) {
+                                    return s[nameProp] === val;
+                                });
+                            return cache[val];
+                        };
+                    }
+
+                    var index;
+
+                    expect(docs[0].$members).toEqual(jasmine.any(Array));
+                    expect(docs[0].$members.length).toBeGreaterThan(5);
+                    index = getIndexer(docs[0], '$members');
+                    expect(index('Code.aStaticMethod')).toBeLessThan(index('Code#bInstanceMethod'));
+                    expect(index('Code#bInstanceMethod')).not.toBeLessThan(index('Code.xStaticMethod'));
+                    expect(index('Code#instanceMethod')).not.toBeLessThan(index('Code.utils'));
+
+                    expect(docs[1].$members).toEqual(jasmine.any(Array));
+                    expect(docs[1].$members.length).toEqual(1); // namespace.location
+
+                    var sym = docs[1].$members[0]; // location symbol
+                    expect(sym.properties.length).toBeGreaterThan(5);
+                    index = getIndexer(sym, 'properties', 'name'); // properties of location
+                    expect(index('href')).toBeLessThan(index('query'));
+                    expect(index('host')).toBeLessThan(index('hostname'));
+                    expect(index('path')).toBeLessThan(index('pathname'));
+                    expect(index('hash')).toBeLessThan(index('host'));
+                    expect(index('origin')).toBeLessThan(index('query'));
+
+                })
+                .catch(function (err) {
+                    expect(Boolean(err)).toEqual(false);
+                    console.log(err.stack || err);
+                })
+                .finally(done);
+        });
+
+        it('should parse source code and output', function (done) {
             options.files = null;
-            options.source = '/**\n *  describe\n *  @namespace\n *  @type {Object}\n */\nconst nspace = {};\nclass Test { constructor() {} }';
+            options.source = JS_SOURCE;
+            options.output = {
+                path: './test/output/docs-from-source.json',
+                force: true,
+                indent: true
+            };
             jsdocx.parse(options)
                 .then(function (docs) {
                     // console.log(docs);
@@ -124,10 +283,14 @@
                     expect(nspace).toBeDefined();
                     var testClass = _.find(docs, { longname: 'Test' });
                     expect(testClass).toBeDefined();
+                    return helper.exists(options.output.path);
+                })
+                .then(function (exists) {
+                    expect(exists).toEqual(true);
                 })
                 .catch(function (err) {
-                    console.log(err);
                     expect(Boolean(err)).toEqual(false);
+                    console.log(err.stack || err);
                 })
                 .finally(done);
         });
@@ -138,27 +301,6 @@
             jsdocx.parse(options)
                 .catch(function (err) {
                     expect(Boolean(err)).toEqual(true);
-                })
-                .finally(done);
-        });
-
-        it('should create output file', function (done) {
-            options.output = {
-                path: './test/output/docs.json',
-                force: true,
-                indent: true
-            };
-            jsdocx.parse(options)
-                .then(function (docs) {
-                    expect(docs).toEqual(jasmine.any(Array));
-                    return helper.exists(options.output.path);
-                })
-                .then(function (exists) {
-                    expect(exists).toEqual(true);
-                })
-                .catch(function (err) {
-                    expect(Boolean(err)).toEqual(false);
-                    console.log(err.stack || err);
                 })
                 .finally(done);
         });